Revert removal and fix MRCompactorAvroKeyDedupJobRunnerTest
