+++ gobblin-compaction/src/test/resources/dedup-schema/dedup-schema-with-pkey.avsc
