+++ gobblin-compaction/src/test/resources/dedup-schema/dedup-schema-without-pkey.avsc
